Implement Dual Bank UART OTA for STM32F4
- Created `EcoflowSTM32F4_Bootloader` project with Dual Bank swap logic.
- Implemented OTA protocol in STM32 Bootloader (Start, Chunk, End, ACK/NACK).
- Implemented Flash Erase/Write logic targeting the Inactive Bank.
- Updated ESP32 `Stm32Serial` to include `otaTask` for streaming firmware from LittleFS.
- Updated STM32 `uart_task` to handle `CMD_OTA_START` and reboot to bootloader.
- Added `bootloader.h` helper for backup register manipulation and system reset.
- Fixed offset calculation logic in ESP32 to track actual bytes sent.
- Added flash verification (memcmp) after each write in bootloader.
- Refactored ESP32 logging to reduce clutter.
- Added visual feedback (Green LED) on successful OTA completion.

+++ EcoflowSTM32F4/src/ui/ui_view_wave2.c
@@ -287,7 +287,7 @@ void ui_view_wave2_update(Wave2DataStruct * data) {
 
     if (lv_slider_is_dragged(arc_set_temp) == false) {
         lv_arc_set_value(arc_set_temp, data->setTemp);
-        lv_label_set_text_fmt(label_set_temp_val, "%d C", data->setTemp);
+        lv_label_set_text_fmt(label_set_temp_val, "%d C", (int)data->setTemp);
     }
 
     if (lv_dropdown_get_selected(dd_sub_mode) != data->subMode) {
